requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
python-dotenv
pytest